[feature] Added the PerformanceReporter (6h)
