+++ lib/abstract_importer/reporters.rb
@@ -1,0 +1,4 @@
+require 'abstract_importer/reporters/base_reporter'
+require 'abstract_importer/reporters/debug_reporter'
+require 'abstract_importer/reporters/null_reporter'
+require 'abstract_importer/reporters/performance_reporter'
