+++ lib/abstract_importer/reporters/null_reporter.rb
@@ -1,0 +1,19 @@
+module AbstractImporter
+  module Reporters
+    class NullReporter < BaseReporter
+      
+      def start_all(importer)
+      end
+      
+      def finish_all(importer, ms)
+      end
+      
+      def finish_setup(ms)
+      end
+      
+      def start_collection(collection)
+      end
+      
+    end
+  end
+end
